Put the magic evseq:// stuff in one place

+++ eventsequencerlib/managedresources.cpp
@@ -2,6 +2,7 @@
 
 #include <QDir>
 #include <QUuid>
+#include <QDebug>
 
 void ManagedResources::mkpathManagedDirectory()
 {
@@ -41,19 +42,32 @@ QString ManagedResources::withSpecifiedName(QString name, QString suffix)
 bool ManagedResources::deleteUrl(QUrl url)
 {
     QString fileName;
-    if (url.scheme() == "evseq" && url.host() == "managed") {
-        if (fileResourceDirectory().isEmpty()) {
-            return false;
-        }
-        fileName = fileResourceDirectory() + url.path();
-    } else {
-        fileName = url.toLocalFile();
-    }
+    if (!convertToFileName(url, &fileName)) return false;
 
     return QFile::remove(fileName);
 }
 
+bool ManagedResources::convertToFileName(QUrl url, QString *fileName)
+{
+    Q_ASSERT(fileName != nullptr);
+    if (url.scheme() == "evseq" && url.host() == "managed") {
+        if (fileResourceDirectory().isEmpty()) {
+            return false;
+        }
+        *fileName = fileResourceDirectory() + url.path();
+    } else {
+        *fileName = url.toLocalFile();
+    }
+    return true;
+}
+
 ManagedResources::ManagedResources(QObject *parent) : QObject(parent)
+{
+
+}
+
+ManagedResources::ManagedResources(QString fileResourceDirectory, QObject *parent)
+    : fileResourceDirectory_(fileResourceDirectory)
 {
 
 }
